api/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
pydantic-settings==2.10.1
PyJWT==2.10.1
cryptography==45.0.5
posthog==6.0.2